+++ frontend/src/components/Ferramentas/produtoseservicos.jsx
@@ -1,0 +1,14 @@
+ import "../../styles/tailwind.css";
+
+
+ const ProdutosEServicos = () => {
+   return ( 
+          <div> 
+            <ul className="ml-2 space-y-2">
+                <li>Produtos & Serviços</li>
+            </ul>
+          </div>  
+        );
+    };
+            
+export default ProdutosEServicos;
