--- frontend/src/components/common/login.jsx
@@ -1,80 +1,0 @@
-import "../../styles/tailwind.css"; // Certifique-se de que as diretivas do Tailwind CSS estão neste arquivo
-
-function Login() {
-  return (
-    <div className="flex items-center justify-center min-h-screen bg-gray-900">
-      <div className="w-full max-w-md p-8 space-y-8 bg-gray-800 rounded-xl shadow-2xl">
-        <h2 className="text-3xl font-bold text-center text-white">Login</h2>
-
-        <form className="space-y-6" action="#" method="POST">
-          <div>
-            <label
-              htmlFor="email"
-              className="block mb-2 text-sm font-medium text-gray-300"
-            >
-              E-mail
-            </label>
-            <input
-              type="email"
-              id="email"
-              className="bg-gray-700 border border-gray-600 text-white text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5"
-              placeholder="seu@email.com"
-              required
-            />
-          </div>
-
-          <div>
-            <label htmlFor="password" className="login-texto-senha">
-              Senha
-            </label>
-            <input
-              type="password"
-              id="password"
-              className="login-campo-senha"
-              placeholder="••••••••"
-              required
-            />
-          </div>
-
-          <div className="flex items-center justify-between">
-            <div className="flex items-center">
-              <input
-                id="remember-me"
-                type="checkbox"
-                className="w-4 h-4 text-blue-600 bg-gray-700 rounded border-gray-600 focus:ring-blue-500 focus:ring-2"
-              />
-              <label
-                htmlFor="remember-me"
-                className="ml-2 text-sm font-medium text-gray-300"
-              >
-                Lembrar de mim
-              </label>
-            </div>
-
-            <a
-              href="#"
-              className="text-sm font-medium text-blue-500 hover:underline"
-            >
-              Esqueceu a senha?
-            </a>
-          </div>
-          <button
-            type="submit"
-            className="w-full text-white bg-blue-600 hover:bg-blue-700 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center"
-          >
-            Entrar
-          </button>
-
-          <p className="text-sm font-light text-gray-400">
-            Ainda não tem uma conta?{" "}
-            <a href="#" className="font-medium text-blue-500 hover:underline">
-              Cadastre-se
-            </a>
-          </p>
-        </form>
-      </div>
-    </div>
-  );
-}
-
-export default Login;
